camera3: Fix print format for ssize_t type
Print format for ssize_t is %zd.
But for some reason is not working,
so explicit cast and print as long int.

+++ camera/QCamera2/HAL3/QCamera3HWI.cpp
@@ -913,23 +913,23 @@ int QCamera3HardwareInterface::validateCaptureRequest(
         QCamera3Channel *channel =
             static_cast<QCamera3Channel*>(b->stream->priv);
         if (channel == NULL) {
-            ALOGE("%s: Request %d: Buffer %d: Unconfigured stream!",
-                    __func__, frameNumber, idx);
+            ALOGE("%s: Request %d: Buffer %ld: Unconfigured stream!",
+                    __func__, frameNumber, (long)idx);
             return BAD_VALUE;
         }
         if (b->status != CAMERA3_BUFFER_STATUS_OK) {
-            ALOGE("%s: Request %d: Buffer %d: Status not OK!",
-                    __func__, frameNumber, idx);
+            ALOGE("%s: Request %d: Buffer %ld: Status not OK!",
+                    __func__, frameNumber, (long)idx);
             return BAD_VALUE;
         }
         if (b->release_fence != -1) {
-            ALOGE("%s: Request %d: Buffer %d: Has a release fence!",
-                    __func__, frameNumber, idx);
+            ALOGE("%s: Request %d: Buffer %ld: Has a release fence!",
+                    __func__, frameNumber, (long)idx);
             return BAD_VALUE;
         }
         if (b->buffer == NULL) {
-            ALOGE("%s: Request %d: Buffer %d: NULL buffer handle!",
-                    __func__, frameNumber, idx);
+            ALOGE("%s: Request %d: Buffer %ld: NULL buffer handle!",
+                    __func__, frameNumber, (long)idx);
             return BAD_VALUE;
         }
     }
@@ -940,23 +940,23 @@ int QCamera3HardwareInterface::validateCaptureRequest(
         QCamera3Channel *channel =
                 static_cast<QCamera3Channel*>(b->stream->priv);
         if (channel == NULL) {
-            ALOGE("%s: Request %d: Buffer %d: Unconfigured stream!",
-                    __func__, frameNumber, idx);
+            ALOGE("%s: Request %d: Buffer %ld: Unconfigured stream!",
+                    __func__, frameNumber, (long)idx);
             return BAD_VALUE;
         }
         if (b->status != CAMERA3_BUFFER_STATUS_OK) {
-            ALOGE("%s: Request %d: Buffer %d: Status not OK!",
-                    __func__, frameNumber, idx);
+            ALOGE("%s: Request %d: Buffer %ld: Status not OK!",
+                    __func__, frameNumber, (long)idx);
             return BAD_VALUE;
         }
         if (b->release_fence != -1) {
-            ALOGE("%s: Request %d: Buffer %d: Has a release fence!",
-                    __func__, frameNumber, idx);
+            ALOGE("%s: Request %d: Buffer %ld: Has a release fence!",
+                    __func__, frameNumber, (long)idx);
             return BAD_VALUE;
         }
         if (b->buffer == NULL) {
-            ALOGE("%s: Request %d: Buffer %d: NULL buffer handle!",
-                    __func__, frameNumber, idx);
+            ALOGE("%s: Request %d: Buffer %ld: NULL buffer handle!",
+                    __func__, frameNumber, (long)idx);
             return BAD_VALUE;
         }
         idx++;
